Add proto2 definition for case classes in scalechain-proto - Why? Need protobuf serialization/deserialization for better performance. - How? 1. Describe common protobuf messages is message-parts.proto. 2. transaction, block message is described in message-parts.proto as they are used by wallet-parts.proto. 3. We will have to convert case classes generated by protobuf to ones that we defined already.
Issue : #93

+++ scalechain-proto/src/main/scala/io/scalechain/blockchain/proto/indexparts.scala
@@ -17,7 +17,7 @@ case class BlockFileInfo(
 ) extends ProtocolMessage
 
 case class BlockInfo(
-  height : Int,
+  height : Long,
   // The total (estimated) number of hash calculations from the genesis block.
   chainWork : Long, // TODO : BUGBUG : Make sure that the 64 bit integer value is enough for the chainwork.
   nextBlockHash : Option[Hash],
